Extract BuildGetterFromSetter to a helper class
to make it reusable from StartConfiguration<TFake>
