+++ src/FakeItEasy.netstd/FakeItEasy.netstd.csproj
@@ -69,6 +69,9 @@
     <Compile Include="..\FakeItEasy\Configuration\IThenConfiguration.cs">
       <Link>Configuration\IThenConfiguration.cs</Link>
     </Compile>
+    <Compile Include="..\FakeItEasy\Configuration\PropertyExpressionHelper.cs">
+      <Link>Configuration\PropertyExpressionHelper.cs</Link>
+    </Compile>
     <Compile Include="..\FakeItEasy\Configuration\PropertySetterConfiguration.cs">
       <Link>Configuration\PropertySetterConfiguration.cs</Link>
     </Compile>
