+++ src/FakeItEasy/Configuration/FakeConfigurationManager.cs
@@ -2,9 +2,7 @@ namespace FakeItEasy.Configuration
 {
     using System;
     using System.Diagnostics.CodeAnalysis;
-    using System.Linq;
     using System.Linq.Expressions;
-    using System.Reflection;
     using FakeItEasy.Core;
     using FakeItEasy.Expressions;
 
@@ -70,38 +68,12 @@ namespace FakeItEasy.Configuration
             this.AssertThatMemberCanBeIntercepted(parsedCallExpression);
 
             var fake = Fake.GetFakeManager(parsedCallExpression.CallTarget);
-            var parsedSetterCallExpression = BuildSetterFromGetter<TValue>(parsedCallExpression);
+            var parsedSetterCallExpression = PropertyExpressionHelper.BuildSetterFromGetter<TValue>(parsedCallExpression);
 
             return new PropertySetterConfiguration<TValue>(
                 parsedSetterCallExpression,
                 newParsedSetterCallExpression =>
                     this.CreateVoidArgumentValidationConfiguration(fake, newParsedSetterCallExpression));
-        }
-
-        private static string GetPropertyName(ParsedCallExpression parsedCallExpression)
-        {
-            var calledMethod = parsedCallExpression.CalledMethod;
-            if (HasThis(calledMethod) && calledMethod.IsSpecialName)
-            {
-                var methodName = calledMethod.Name;
-                if (methodName.StartsWith("get_", StringComparison.Ordinal))
-                {
-                    return methodName.Substring(4);
-                }
-            }
-
-            return null;
-        }
-
-        private static bool HasThis(MethodInfo methodCall)
-        {
-            return methodCall.CallingConvention.HasFlag(CallingConventions.HasThis);
-        }
-
-        private static Expression BuildArgumentThatMatchesAnything<TValue>()
-        {
-            Expression<Func<TValue>> value = () => A<TValue>.Ignored;
-            return value.Body;
         }
 
         private static void GuardAgainstNonFake(object target)
@@ -110,62 +82,6 @@ namespace FakeItEasy.Configuration
             {
                 Fake.GetFakeManager(target);
             }
-        }
-
-        private static string GetExpressionDescription(ParsedCallExpression parsedCallExpression)
-        {
-            var matcher = new ExpressionCallMatcher(
-                parsedCallExpression,
-                ServiceLocator.Current.Resolve<ExpressionArgumentConstraintFactory>(),
-                ServiceLocator.Current.Resolve<MethodInfoManager>());
-
-            return matcher.DescriptionOfMatchingCall;
-        }
-
-        private static ParsedCallExpression BuildSetterFromGetter<TValue>(
-            ParsedCallExpression parsedCallExpression)
-        {
-            var propertyName = GetPropertyName(parsedCallExpression);
-            if (propertyName == null)
-            {
-                var expressionDescription = GetExpressionDescription(parsedCallExpression);
-                throw new ArgumentException("Expression '" + expressionDescription +
-                                            "' must refer to a property or indexer getter, but doesn't.");
-            }
-
-            var parsedArgumentExpressions = parsedCallExpression.ArgumentsExpressions ?? new ParsedArgumentExpression[0];
-            var parameterTypes = parsedArgumentExpressions
-                .Select(p => p.ArgumentInformation.ParameterType)
-                .Concat(new[] { parsedCallExpression.CalledMethod.ReturnType })
-                .ToArray();
-
-            var indexerSetterInfo = parsedCallExpression.CallTarget.GetType()
-                .GetMethod("set_" + propertyName, parameterTypes);
-
-            if (indexerSetterInfo == null)
-            {
-                if (parsedArgumentExpressions.Any())
-                {
-                    var expressionDescription = GetExpressionDescription(parsedCallExpression);
-                    throw new ArgumentException("Expression '" + expressionDescription +
-                                                "' refers to an indexed property that does not have a setter.");
-                }
-
-                throw new ArgumentException(
-                    "The property '" + propertyName + "' does not have a setter.");
-            }
-
-            var originalParameterInfos = indexerSetterInfo.GetParameters();
-
-            var newParsedSetterValueExpression = new ParsedArgumentExpression(
-                BuildArgumentThatMatchesAnything<TValue>(),
-                originalParameterInfos.Last());
-
-            var arguments = parsedArgumentExpressions
-                .Take(originalParameterInfos.Length - 1)
-                .Concat(new[] { newParsedSetterValueExpression });
-
-            return new ParsedCallExpression(indexerSetterInfo, parsedCallExpression.CallTarget, arguments);
         }
 
         private IVoidArgumentValidationConfiguration CreateVoidArgumentValidationConfiguration(FakeManager fake, ParsedCallExpression parsedCallExpression)
