+++ src/FakeItEasy/FakeItEasy.csproj
@@ -62,6 +62,7 @@
   <ItemGroup>
     <Compile Include="AsyncReturnValueConfigurationExtensions.cs" />
     <Compile Include="Compatibility\MethodInfoWrapper.cs" />
+    <Compile Include="Configuration\PropertyExpressionHelper.cs" />
     <Compile Include="LinkedListExtensions.cs" />
     <Compile Include="Configuration\IDoNothingConfiguration.cs" />
     <Compile Include="Configuration\IPropertySetterAnyValueConfiguration.cs" />
